+++ src/main/java/roomescape/auth/dto/LoginMember.java
@@ -1,6 +1,6 @@
 package roomescape.auth.dto;
 
-import roomescape.member.MemberRole;
+import roomescape.member.domain.MemberRole;
 
 public record LoginMember(String name, String email, MemberRole role) {
 }
