+++ src/main/java/roomescape/member/domain/Member.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
