+++ src/main/java/roomescape/member/domain/MemberRole.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.domain;
 
 public enum MemberRole{
     ADMIN, MEMBER
