+++ src/main/java/roomescape/member/repository/MemberJpaRepository.java
@@ -1,7 +1,8 @@
-package roomescape.member;
+package roomescape.member.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.member.domain.Member;
 
 public interface MemberJpaRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
