+++ src/main/java/roomescape/member/repository/MemberRepository.java
@@ -1,7 +1,8 @@
-package roomescape.member;
+package roomescape.member.repository;
 
 import java.util.List;
 import java.util.Optional;
+import roomescape.member.domain.Member;
 
 public interface MemberRepository {
