+++ src/main/java/roomescape/reservation/dto/WaitingRankReservation.java
@@ -1,7 +1,8 @@
-package roomescape.reservation;
+package roomescape.reservation.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import roomescape.reservation.domain.Reservation;
 
 @Getter
 @AllArgsConstructor
