+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -1,10 +1,14 @@
-package roomescape.reservation;
+package roomescape.reservation.repository;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.member.Member;
-import roomescape.reservationtime.ReservationTime;
-import roomescape.theme.Theme;
+import roomescape.member.domain.Member;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.ReservationDate;
+import roomescape.reservation.domain.ReservationStatus;
+import roomescape.reservation.dto.WaitingRankReservation;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
 
 public interface ReservationRepository {
     Reservation save(Reservation reservation);
