+++ src/main/java/roomescape/reservation/repository/ReservationRepositoryImpl.java
@@ -1,12 +1,16 @@
-package roomescape.reservation;
+package roomescape.reservation.repository;
 
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import roomescape.member.Member;
-import roomescape.reservationtime.ReservationTime;
-import roomescape.theme.Theme;
+import roomescape.member.domain.Member;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.ReservationDate;
+import roomescape.reservation.domain.ReservationStatus;
+import roomescape.reservation.dto.WaitingRankReservation;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
 
 @Repository
 @RequiredArgsConstructor
