+++ src/main/java/roomescape/reservationtime/domain/ReservationTime.java
@@ -1,10 +1,11 @@
-package roomescape.theme;
+package roomescape.reservationtime.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.time.LocalTime;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -13,27 +14,24 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
-public class Theme {
+public class ReservationTime {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @EqualsAndHashCode.Include
     private final Long id;
 
     @Column(nullable = false)
-    private String name;
+    private final LocalTime startAt;
 
-    @Column(nullable = false)
-    private String description;
-
-    @Column(nullable = false)
-    private String thumbnail;
-
-    protected Theme() {
-        id = null;
+    protected ReservationTime() {
+        this(null, null);
     }
 
-    public Theme(final String name, final String description, final String thumbnail) {
-        this(null, name, description, thumbnail);
+    public ReservationTime(final LocalTime startAt) {
+        this(null, startAt);
+    }
+
+    public boolean isBefore(final LocalTime localTime) {
+        return startAt.isBefore(localTime);
     }
 }
