+++ src/main/java/roomescape/reservationtime/dto/AvailableReservationTimeResponse.java
@@ -3,7 +3,7 @@ package roomescape.reservationtime.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import roomescape.exception.custom.reason.ResponseInvalidException;
-import roomescape.reservationtime.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 
 public record AvailableReservationTimeResponse(
         Long id,
