+++ src/main/java/roomescape/reservationtime/dto/ReservationTimeResponse.java
@@ -2,7 +2,7 @@ package roomescape.reservationtime.dto;
 
 import java.time.LocalTime;
 import roomescape.exception.custom.reason.ResponseInvalidException;
-import roomescape.reservationtime.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 
 public record ReservationTimeResponse(
         Long id,
