+++ src/main/java/roomescape/reservationtime/repository/ReservationTimeRepository.java
@@ -1,8 +1,9 @@
-package roomescape.reservationtime;
+package roomescape.reservationtime.repository;
 
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
+import roomescape.reservationtime.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
     ReservationTime save(ReservationTime reservationTime);
