+++ src/main/java/roomescape/theme/ThemeService.java
@@ -6,9 +6,11 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.exception.custom.reason.theme.ThemeNotFoundException;
 import roomescape.exception.custom.reason.theme.ThemeUsedException;
-import roomescape.reservation.ReservationJpaRepository;
+import roomescape.reservation.repository.ReservationJpaRepository;
+import roomescape.theme.domain.Theme;
 import roomescape.theme.dto.ThemeRequest;
 import roomescape.theme.dto.ThemeResponse;
+import roomescape.theme.repository.ThemeRepository;
 
 @Service
 @AllArgsConstructor
