+++ src/main/java/roomescape/theme/domain/Theme.java
@@ -1,11 +1,10 @@
-package roomescape.reservationtime;
+package roomescape.theme.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import java.time.LocalTime;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -14,24 +13,27 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
-public class ReservationTime {
+public class Theme {
 
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @EqualsAndHashCode.Include
     private final Long id;
 
     @Column(nullable = false)
-    private final LocalTime startAt;
+    private String name;
 
-    protected ReservationTime() {
-        this(null, null);
+    @Column(nullable = false)
+    private String description;
+
+    @Column(nullable = false)
+    private String thumbnail;
+
+    protected Theme() {
+        id = null;
     }
 
-    public ReservationTime(final LocalTime startAt) {
-        this(null, startAt);
-    }
-
-    public boolean isBefore(final LocalTime localTime) {
-        return startAt.isBefore(localTime);
+    public Theme(final String name, final String description, final String thumbnail) {
+        this(null, name, description, thumbnail);
     }
 }
