+++ src/main/java/roomescape/theme/repository/ThemeRepository.java
@@ -1,8 +1,9 @@
-package roomescape.theme;
+package roomescape.theme.repository;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import roomescape.theme.domain.Theme;
 
 public interface ThemeRepository {
     Theme save(Theme theme);
