+++ src/test/java/roomescape/reservationtime/ReservationTimeTest.java
@@ -8,8 +8,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import roomescape.reservationtime.domain.ReservationTime;
 
-public class ReservationTimeTest {
+class ReservationTimeTest {
 
 
     @DisplayName("isBefore 테스트")
